pyproject.toml
[tool.poetry]
name = "gladiator"
version = "0.1.0"
description = "Generate type-safe OpenGL wrappers for C++"
authors = ["hellcat17 <dodgehellcat17@outlook.com>"]
packages = [{ include = "gladiator" }]
include = ["resources/data/**", "resources/templates/**"]

[tool.poetry.dependencies]
python = "^3.9"
attrs = "^20.3"
pyyaml = "^5.3"
ConfigArgParse = "^1.5.3"
Jinja2 = "^3.0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"
black = "20.8b1"
mypy = "0.790"
coverage = "^5.3"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
